+++ src/components/Footer/Footer.styles.jsx
@@ -5,7 +5,7 @@ export const $Footer = styled.div`
     display: flex; 
     justify-content: center; 
     align-items: center; 
-    padding: 20px; 
+    padding: 10px; 
     font-size: 12px; 
     color: ${appwhite}; 
     background-color: ${appblack};